feat: Add a "stopping" time
Currently as soon as you call app.stop(), we almost immendiatly close the servers and drain the connection pools. This can cause any inflight requests to fail. This PR introduces a shutdown timer. After stop is called, the app will start responding as not ready so no new requests should be sent, but we await default 20 seconds before we shut everything down.